update to tdlib v1.7.9 master 8d7bda0

+++ lib/tgchat/ext/td/td/telegram/net/NetQueryStats.h
@@ -1,5 +1,5 @@
 //
-// Copyright Aliaksei Levin (levlam@telegram.org), Arseny Smirnov (arseny30@gmail.com) 2014-2020
+// Copyright Aliaksei Levin (levlam@telegram.org), Arseny Smirnov (arseny30@gmail.com) 2014-2021
 //
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -17,7 +17,7 @@ namespace td {
 
 struct NetQueryDebug {
   double start_timestamp_ = 0;
-  int32 my_id_ = 0;
+  int64 my_id_ = 0;
   int32 resend_count_ = 0;
   string state_ = "empty";
   double state_timestamp_ = 0;
